feat：补充还款详情优惠券试算，新增还款h5页面，补充付款逻辑；修改部分UI

+++ lib/pages/loan/components/loan_order_detail_cards.dart
@@ -215,7 +215,7 @@ class _DetailRowData {
       label: label,
       value: value == null
           ? ''
-          : value.toDouble().formatAmount(showCurrencySymbol: true),
+          : value.formatAmount(showCurrencySymbol: true),
     );
   }
